[Refactor] 패키지 구조 변경
artifact는 chess로 한다.
pawn은 piece패키지 하단에 위치한다.
